requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
pydantic-settings==2.8.0
python-multipart==0.0.20
email-validator==2.2.0
